requirements.txt
flask
flask-restful
scikit-learn
pandas
numpy
psycopg2-binary